Allwinner: remove workaround that for MPEG4 codec
there is an ugly workaround used that for MPEG4 codec the VOL header data is passed sideways to the image data via an newly created VDPAU function call. The whole concept can lead to a race condition and requires changes on the libvdpau and vdpau interface.
This sideway has been removed by passing the VOL header structure together with the mpeg4 video data and the vdpau driver decodes the VOL header by itself. This way the potential race condition can be removed and libvdpau changes are not required (for this functionality). Still, ffmpeg needs a fix to pass the VOL header when it does not exist in the video codec stream (which can happen for some video containers, like .MP4)